Deprecate GenericTypeParameterReferenceString
This commit follows the recommendation in the DocBit of the concept to
make the GenericTypeParameterReference a TypeReference. The scope for
the reference is determined using the ScopeProvider mechanism, added
to IGenericTypeList. The substitute menu for TypeParameter was removed
as it does not need to be substituted manually.

A migration (with test) is also included to replace string references,
if a type parameter with that name can be resolved.