Fix and simplify type parameter variance
Fixes the issue that CsBaseLanguage only supported variance keywords in
(generic) delegates, while the 5.0 language specification also allows
them in interfaces. The "isVarianceAnnotatable" property of a
TypeParameter now queries the enclosing IGenericTypeList for this
information, so it no longer needs to be managed from code. Two editor
tests have been added to check this feature.