Improve variable editors
The main change is the addition of an optional initializer expression to
the VariableDeclaration. The initializer can be easily added and removed
via a transformation menu entry and action map. This obviates the need
for the LocalVariableDeclarationWithInitializer and the expression node
of the ConstantDeclarator, which have been marked as deprecated and
given appropriate migrations to the updated VariableDeclaration.
Additionally, usages of IVariableDeclarator have been replaced by
VariableDeclarations.

The result is a smoother editing experience for variable declarations
and related concepts, similar to that of text editors.

A side effect is that const declarations can be left uninitialized,
which is marked as an error by a checking rule. This rule is validated
by a test, as are the other changes.