Implement await expressions
Await expressions are constrained to require an ancestor of the new
interface ICanBeAsync, which applies to methods and anonymous functions
(yet to be implemented). For a better editing experience, the other
constraints in the specification are instead handled by checking rules.
An intention is provided to add the async modifier to the enclosing
function if this is required, similar to the one found in C# IDEs.