Improve handling of generic parameters
The AddGenericParameterListSection transformation menu is now only
included if the referenced type is generic, which is inferred using a
new method. A descriptive text has been added to the transformation
menus for generic type lists and some changes were made to their editors
to remove extra spaces from the textgen output.